Move db delete logic to RecordDeleteViewModel
